requirements.txt
requests>=2.0.0
python-dotenv>=0.15.0